Fix gradle wrapper: add missing files.
When using the Gradle wrapper (executed with ./gradlew) instead of a local
Gradle install, the gradle-wrapper.jar needs to be present.

This commit removes the '/gradle' .gitignore entry since it isn't a build
artifact path. It also adds the missing gradle-wrapper.jar and a
gradle-wrapper.properties configuration file.
The distribution to be used by the wrapper is set to 'gradle-5.6.4-bin', which
is the same version used by the CI.